revert moving indexer api folder
