+++ frontend/codegen.ts
@@ -5,7 +5,7 @@ const config: CodegenConfig = {
   documents: ['src/**/*.{ts,tsx}'],
   ignoreNoDocuments: true, // for better experience with the watcher
   generates: {
-    './src/api/graphql/': {
+    './src/features/history/graphql/': {
       preset: 'client',
       plugins: [],
       config: {
