--- frontend/src/api/consts.ts
@@ -1,3 +1,0 @@
-const INDEXER_ADDRESS = import.meta.env.VITE_INDEXER_ADDRESS as string;
-
-export { INDEXER_ADDRESS };
