+++ frontend/src/features/history/hooks/index.ts
@@ -1,5 +1,6 @@
+import { usePairs } from './use-pairs';
 import { useTransactionFilters } from './use-transaction-filters';
 import { useTransactions } from './use-transactions';
 import { useTransactionsCount } from './use-transactions-count';
 
-export { useTransactionsCount, useTransactions, useTransactionFilters };
+export { usePairs, useTransactionsCount, useTransactions, useTransactionFilters };
