+++ frontend/src/features/history/index.ts
@@ -1,6 +1,6 @@
 import { List, TransactionsCounter, TransactionCard, PendingTransactionsWarning } from './components';
 import { FIELD_NAME, STATUS_OPTIONS, TIMESTAMP_OPTIONS, TRANSACTIONS_LIMIT } from './consts';
-import { useTransactions, useTransactionsCount, useTransactionFilters } from './hooks';
+import { usePairs, useTransactions, useTransactionsCount, useTransactionFilters } from './hooks';
 import { getAssetOptions } from './utils';
 
 export {
@@ -8,6 +8,7 @@ export {
   TransactionsCounter,
   TransactionCard,
   PendingTransactionsWarning,
+  usePairs,
   useTransactions,
   useTransactionsCount,
   useTransactionFilters,
